[GridFragmentation] Ensure we layout all items when size is saturated.
Previously it was possible to saturate the size of the grid, with some
content, plus a LayoutUnit::Max() end border.

In this particular sub-case - we'd believe that we were at the end of
content, then make the first fragment LayoutUnit::Max() (with the
border).

The PlaceGridItemsForFragmentation function looks at the previously
consumed block-size to place items.

For this case, we'd layout some content in the first fragment.
Then in the second fragment believe we were past all the content due
to the value of ConsumedBlockSize().

The fix for this is to separately track the "consumed_grid_size" to
ensure the end border doesn't mess up (saturate) this value used for
positioning.

Bug: 1319964
Change-Id: I8cc98ab61d7148cf64c691c4322dca7a68c5fe95